requirements.txt
itaxotools-spart-parser @ git+https://github.com/iTaxoTools/SpartParser.git@main
shapely
geopy
